Move random IP address generators to TestUtils library. Could be useful for generating benchmarks and stuff.
